Perguntas/requirements-dev.txt
## Dev/Test requirements
## Inclui base via requirements.txt para garantir dependências (mysqlclient etc.)
-r requirements.txt

pytest==8.3.2
pytest-django==4.8.0
pytest-cov>=5.0.0
ruff>=0.6.0

